back-model : DAO 연결 및 app 실행 확인
