+++ .gitignore
@@ -35,7 +35,9 @@ share/python-wheels/
 *.egg-info/
 .installed.cfg
 *.egg
+.config.py
 MANIFEST
+
 
 # PyInstaller
 #  Usually these files are written by a python script from a template
