+++ manage.py
@@ -1,17 +1,7 @@
-import sys
 
-from flask_script   import Manager
-from .app            import create_app
-from flask_twisted  import Twisted
-from twisted.python import log
+from app           import create_app
 
 if __name__ == "__main__":
     app = create_app()
 
-    twisted = Twisted(app)
-    log.startLogging(sys.stdout)
-
-    app.logger.info(f"Running the app...")
-
-    manager = Manager(app)
-    manager.run()
+    app.run("localhost", 5000, debug=True)
