--- model/VO/__init__.py
@@ -1,26 +1,0 @@
-
-from .Category import Category
-from .Click_data import Click_data
-from .Image import Image
-from .Likes import Likes
-from .Major import Major
-from .Middle import Middle
-from .R_image import R_image
-from .Rec_tag import Rec_tag
-from .Tag import Tag
-from .Tag_pf import Tag_pf
-from .User import User
-
-__all__ = [
-    'Category',
-    'Click_data',
-    'Image',
-    'Likes',
-    'Major',
-    'Middle',
-    'Rec_tag',
-    'R_image',
-    'Tag',
-    'Tag_pf',
-    'User'
-]
