+++ model/__init__.py
@@ -1,11 +1,8 @@
-from flask_sqlalchemy import SQLAlchemy
 
-from .image_dao import ImageDAO
-from .tag_dao import TagDAO
-from .user_dao import UserDAO
-from .util import query2list
-
-db = SQLAlchemy()
+from model.image_dao import ImageDAO
+from model.tag_dao import TagDAO
+from model.user_dao import UserDAO
+from model.util import query2list
