+++ model/tag_dao.py
@@ -1,5 +1,4 @@
 from sqlalchemy import text
-from model.VO import Tag
 
 class TagDAO:
     def __init__(self, database):
